[code coverage] Implement Chromium code coverage plugin
This CL implements the Chromium code coverage plugin that repeatedly
makes requests (in an exponential backoff fashion) to a service to
fetch code coverage data for the currently shown change/patchset, and
once the data is ready, it uses the annotation layer APIs to highlight
the uncovered lines.

In order to control the possible negative impact when something is
wrong (e.g. a bug that causes the wrong lines to be highlighted), a
switch will be implemented at the service side that allows anyone with
a google@ account to turn of the service with just one click.

There is a bug with the annotation APIs that causes this plugin to fail
whenever an user refreshes the page when they are at the diff view
page: https://bugs.chromium.org/p/gerrit/issues/detail?id=10054, it
will be investigated and fixed separately.

Bug: gerrit:10067
Change-Id: Id4f95fba4f5eeffec39abf11ddd31cf322ae1783