[code coverage] Decrease start sleep time
There was a racing condition that when the user clicks a file or
'Expand All' too fast, the coverage data won't be shown.

This CL fixes the issue by decreasing the start sleep time to make the
plugin attempts fetching coverage data more frequently at first, and
given that the exponential backoff behavior is maintained, the
increased overhead is negligible.

Bug: 917152
Change-Id: I4e0b0d5639ba4d8dd50cf84896d583aa9fee7549