[code coverage] Clean up legacy buggy code
The removed code was added to work handle the issue when user switches
between changes or patchsets, now that the issue is handled by the
"showchange" callback more gracefully, the work around is not needed
anymore, and what's more, due to that the same issue is unnecessarily
handled twice, they cause a bug where annotations may disapear
unexpectedly when switching between patchsets.

This CL fixes the issue by removing the unnecessary workaround.

Bug: 915856
Change-Id: I29ee76f4b6e3588f6cb5251b209f4eb53c126793