[code coverage] Fix coverage columns styling
According to Polymer documentation: "Styles do not leak into the local
DOM from above, and styles do not leak from one local DOM into the
local DOM of other elements inside.".

Which results in that the global styling fails to apply to the coverage
column elements, and this CL fixes the issue.

Bug: 939904
Change-Id: Ib833eadab96c80fe8c0dde499367fd69cde58c9c