Allow overriding default Chromium coverage endpoints
Enable the use of the code-coverage plugin with a non-Chromium
project, allowing to define an arbitrary endpoint in the
plugins' configuration on a per-project basis.

Allows also to work in a local development mode where the
code-coverage data can be stored locally on the $GERRIT_SITE/static
folder and served to the GUI through Gerrit's static assets
servlet.

See [1] for a showcase on how a local test environment is
possible by leveraging this change.

For existing setups and configs, all the current endpoints
URLs are honoured as before.

[1] https://imgur.com/a/Rv9ThnP

Change-Id: I2eaaaee7b9a130909ff882f98a2cb9f7ecdc6e45