Allow overriding default Chromium converage endpoints
Enable the use of the code-coverage plugin with a non-Chromium
project, allowing to define an arbitrary endpoint in the
plugins' configuration on a per-project basis.

Allows also to work in a local development mode where the
code-coverage data can be stored locally on the $GERRIT_SITE/static
folder and served to the GUI through Gerrit's static assets
servlet.

Change-Id: I2eaaaee7b9a130909ff882f98a2cb9f7ecdc6e45